chore: 기존에 존재했던 인수테스트 코드와 새로 만든 인수테스트 코드 분리
