+++ src/test/java/cart/integration/IntegrationTest.java
@@ -1,19 +1,56 @@
 package cart.integration;
 
-import io.restassured.RestAssured;
-import org.junit.jupiter.api.BeforeEach;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.web.server.LocalServerPort;
-import org.springframework.test.context.jdbc.Sql;
+import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 
-@Sql("/init-integration-data.sql")
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class IntegrationTest {
-    @LocalServerPort
-    private int port;
+import cart.dao.CartItemDao;
+import cart.dao.MemberDao;
+import cart.repository.CouponRepository;
+import cart.repository.MemberCouponRepository;
+import cart.repository.OrderRepository;
+import cart.repository.ProductRepository;
+import org.junit.jupiter.api.BeforeEach;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.setup.MockMvcBuilders;
+import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.context.WebApplicationContext;
+import org.springframework.web.filter.CharacterEncodingFilter;
+
+@SpringBootTest
+@AutoConfigureMockMvc
+@Transactional
+public abstract class IntegrationTest {
+
+    protected MockMvc mockMvc;
+
+    @Autowired
+    protected WebApplicationContext ctx;
+
+    @Autowired
+    protected MemberDao memberdao;
+
+    @Autowired
+    protected ProductRepository productRepository;
+
+    @Autowired
+    protected CartItemDao cartItemDao;
+
+    @Autowired
+    protected OrderRepository orderRepository;
+
+    @Autowired
+    protected CouponRepository couponRepository;
+
+    @Autowired
+    protected MemberCouponRepository memberCouponRepository;
 
     @BeforeEach
-    void setUp() {
-        RestAssured.port = port;
+    void setMockMvc() {
+        this.mockMvc = MockMvcBuilders.webAppContextSetup(ctx)
+                .addFilters(new CharacterEncodingFilter("UTF-8", true))  // 필터 추가
+                .alwaysDo(print())
+                .build();
     }
 }
